requirements.txt
streamlit>=1.18.0
google-generativeai
langchain
langchain-google-genai
langchain-huggingface
langchain-community
pypdf
faiss-cpu
sentence-transformers
python-dotenv
huggingface_hub
chromadb